+++ CHANGES.md
@@ -1,4 +1,7 @@
-# v1.0.0
+# v0.0.2
+- Add `ready` orders tab
+- Drivers can grab `ready` orders
 
+# v0.0.1
 - Backend integration
 - Login Driver